Added TeamSlot script, and added game objects for the enemy's team and player's team
