+++ Assets/TeamSlot.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0637b0b3721bd174fb9ad3d0b24b0f7f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
